Initial Push
Currently, scoring is based off of time, and then every 10 seconds an enemy spawns (with an additional enemy for every ten seconds that have passed) So at 1 minute you'd have 6 enemies. This is impossible to beat and very aggravating, so I'll switch to a new enemy being spawned every time an enemy dies.